// frontend/src/enums/enums.tsx
/* 
This module is for global enums.
In general keep interfaces/enums as close to their useage.
*/

export enum SortDirection {
  ASCENDING = "asc",
  DESCENDING = "desc",
}

export enum Race {
  ZERG = "zerg",
  TERRAN = "terran",
  PROTOSS = "protoss",
  RANDOM = "random",
}

export enum MatchUp {
  ZVT = "ZVT",
  ZVP = "ZVP",
  ZVZ = "ZVZ",
  ZVR = "ZVR",
  TVT = "TVT",
  TVP = "TVP",
  TVZ = "TVZ",
  TVR = "TVR",
  PVT = "PVT",
  PVP = "PVP",
  PVZ = "PVZ",
  PVR = "PVR",
  RVT = "RVT",
  RVP = "RVP",
  RVZ = "RVZ",
  RVR = "RVR",
}